pyproject.toml
[tool.poetry]
name = "trend-news-notifier"
version = "0.1.0"
description = ""
authors = ["nguyennampfiev <nguyennampfiev1995@gmail.com>"]
readme = "README.md"
packages = [{include = "news_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
pydantic = {extras = ["email"], version = "^2.12.0"}
python-dotenv = "^0.21.0"
aiohttp = "^3.12.15"
beautifulsoup4 = "^4.13.4"
duckduckgo-search = "^8.1.1"
ddgs = "^9.5.4"
pre-commit = "^4.3.0"
openai-agents = "^0.2.9"
aiosmtplib = "^4.0.2"
sqlalchemy = "^2.0.44"
aiosqlite = "^0.21.0"
numpy = "^2.3.3"
black = "^25.9.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
pre-commit = "^4.3.0"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
aioresponses = "^0.7.8"
black = "^25.1.0"
isort = "^6.0.1"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
